Red_4 lista ahora si

+++ Algoritmo/Potencias/Red_4/c_multiclass_red_neuronal_tesis.py
@@ -19,7 +19,12 @@ from itertools import product
 # =============================================================================
 # Importing the dataset
 # =============================================================================
-a = list(product([2500,350,500,1000,2500,5000,10000],['s','s']))
+a = list(product([30,50,100,250,350,500,1000,2500,5000],['s','n']))
+duration = 1 #segundos
+f1 = 440    #término de procesos simples
+f2 = 550    #Termino de grid search
+f3 = 650    #Termino de cross_validation
+f4 = 750    #Termino del codigo, paso al siguiente
 
 for i in range(len(a)):
   
@@ -28,9 +33,6 @@ for i in range(len(a)):
     print(f"j = {a[i][0]}\nson = {a[i][1]}")
 
     directory = os.getcwd()
-
-    if j==1000 and son=='S' or j==1000 and son=='N':
-        continue
     
     if os.path.isdir(os.getcwd() + '/{}_{}'.format(j,son)) == True:
         pass
@@ -39,41 +41,47 @@ for i in range(len(a)):
     
     t = TicToc()
     
-    df1 = fixrows('Potencia_r1').iloc[:j,:]
-    num_row = np.shape(df1)[0]
-    coords  = ['(1,0)' for j in range(num_row)]
-    coords = pd.DataFrame(coords,dtype=object, columns = ['X,Y'])
-    df1 = coords.join(df1, how='left')
-    df1.to_csv('Potencia_R1.csv')
-     
-    df2 = fixrows('Potencia_r2').iloc[:j,:]
-    num_row = np.shape(df2)[0]
-    coords  = ['(0,0)' for j in range(num_row)]
-    coords = pd.DataFrame(coords,dtype=object, columns = ['X,Y'])
-    df2 = coords.join(df2, how='left')
-    df2.to_csv('Potencia_R2.csv')
-     
-    df3 = fixrows('Potencia_r3').iloc[:j,:]
-    num_row = np.shape(df3)[0]
-    coords  = ['(0,1)' for j in range(num_row)]
-    coords = pd.DataFrame(coords,dtype=object, columns = ['X,Y'])
-    df3 = coords.join(df3, how='left')
-    df3.to_csv('Potencia_R3.csv')
-    
-    df4 = fixrows('Potencia_r4').iloc[:j,:]
-    num_row = np.shape(df4)[0]
-    coords  = ['(1,1)' for j in range(num_row)]
-    coords = pd.DataFrame(coords,dtype=object, columns = ['X,Y'])
-    df4 = coords.join(df4, how='left')
-    df4.to_csv('Potencia_R4.csv')
-      
-    # Fusionar archivos corregidos para obtener el archivo de potencias final
-    fusionar_csv('Potencia_R1','Potencia_R2','Potencia_R3','Potencia_R4')
-     
-    df0 = fixrows('potencias_fusionado').iloc[:,1:]
-    os.system('rm Potencia_R*')
-    os.system('rm Potencia_r*_corregido.csv')
-    df0 = pd.read_csv('potencias_fusionado_corregido.csv').iloc[3:,1:]
+    if son=='S':
+
+        df1 = fixrows('Potencia_r1').iloc[:j,:]
+        num_row = np.shape(df1)[0]
+        coords  = ['(1,0)' for j in range(num_row)]
+        coords = pd.DataFrame(coords,dtype=object, columns = ['X,Y'])
+        df1 = coords.join(df1, how='left')
+        df1.to_csv('Potencia_R1.csv')
+         
+        df2 = fixrows('Potencia_r2').iloc[:j,:]
+        num_row = np.shape(df2)[0]
+        coords  = ['(0,0)' for j in range(num_row)]
+        coords = pd.DataFrame(coords,dtype=object, columns = ['X,Y'])
+        df2 = coords.join(df2, how='left')
+        df2.to_csv('Potencia_R2.csv')
+         
+        df3 = fixrows('Potencia_r3').iloc[:j,:]
+        num_row = np.shape(df3)[0]
+        coords  = ['(0,1)' for j in range(num_row)]
+        coords = pd.DataFrame(coords,dtype=object, columns = ['X,Y'])
+        df3 = coords.join(df3, how='left')
+        df3.to_csv('Potencia_R3.csv')
+        
+        df4 = fixrows('Potencia_r4').iloc[:j,:]
+        num_row = np.shape(df4)[0]
+        coords  = ['(1,1)' for j in range(num_row)]
+        coords = pd.DataFrame(coords,dtype=object, columns = ['X,Y'])
+        df4 = coords.join(df4, how='left')
+        df4.to_csv('Potencia_R4.csv')
+          
+        # Fusionar archivos corregidos para obtener el archivo de potencias final
+        fusionar_csv('Potencia_R1','Potencia_R2','Potencia_R3','Potencia_R4')
+         
+        df0 = fixrows('potencias_fusionado').iloc[:,1:]
+        os.system('rm Potencia_R*')
+        os.system('rm Potencia_r*_corregido.csv')
+        df0 = pd.read_csv('potencias_fusionado_corregido.csv').iloc[3:,1:]
+        os.system('play -nq -t alsa synth {} sine {}'.format(duration,f1))
+    else:
+        df0 = pd.read_csv('potencias_fusionado_corregido.csv').iloc[3:,1:]
+        os.system('play -nq -t alsa synth {} sine {}'.format(duration,f1))
     
     X = df0.iloc[:,1:].values #variables Dependientes (Potencias)
     y = df0.iloc[:,0].values #values Independientes (Posición)
@@ -207,9 +215,35 @@ for i in range(len(a)):
     #a = optimizer, b= epochs,c=batch_size
 
     if j==30:
-    	best_parameters={'optimizer': 'adam', 'epochs': 40, 'batch_size': 16}
-    elif j==50 or j==100 or j==250 or j==350 or j==500:
-    	best_parameters={'optimizer': 'rmsprop', 'epochs': 25, 'batch_size': 48}
+        if son=='N':
+            best_parameters={'optimizer': 'adam', 'epochs': 40, 'batch_size': 16}
+        else:
+            best_parameters={'optimizer': 'adam', 'epochs': 35, 'batch_size': 16}
+    elif j==50:
+        if son=='N':
+            best_parameters={'optimizer': 'rmsprop', 'epochs': 25, 'batch_size': 16}
+        else:
+            best_parameters={'optimizer': 'rmsprop', 'epochs': 35, 'batch_size': 16}
+    elif j==100:
+        if son=='N':
+            best_parameters={'batch_size': 16, 'epochs': 35, 'optimizer': 'adam'}
+        else:
+            best_parameters={'batch_size': 16, 'epochs': 25, 'optimizer': 'adam'}
+    elif j==250:
+        if son=='N':
+            best_parameters={'batch_size': 32, 'epochs': 25, 'optimizer': 'adam'}
+        else:
+            best_parameters={'batch_size': 16, 'epochs': 25, 'optimizer': 'adam'}
+    elif j==350:
+        if son=='N':
+            best_parameters={'batch_size': 16, 'epochs': 15, 'optimizer': 'adamax'}
+        else:
+            best_parameters={'batch_size': 16, 'epochs': 25, 'optimizer': 'rmsprop'}
+    elif j==500:
+        if son=='N':
+            best_parameters={'batch_size': 16, 'epochs': 35, 'optimizer': 'adamax'}
+        else:
+            best_parameters={'batch_size': 48, 'epochs': 35, 'optimizer': 'adam'}
     elif j==1000:
     	if son=='N':
             best_parameters={'batch_size': 16, 'epochs': 15, 'optimizer': 'adamax'}
@@ -256,6 +290,7 @@ for i in range(len(a)):
     # save model and architecture to single file
     classifier.save(directory + "/{}_{}/model_{}.h5".format(j,son,j))
     print("Saved model to disk\n")
+    os.system('play -nq -t alsa synth {} sine {}'.format(duration,f3))
     
     # =============================================================================
     #     Escritura de archivo
@@ -317,3 +352,4 @@ for i in range(len(a)):
     os.system('mv Accuracy.png '+ mv + 'Accuracy.png')
     os.system('mv Classification_report.csv ' + mv + 'Classification_report.csv')
     os.system('mv model.json ' + mv + 'model_{}_{}.json'.format(j,son))
+    os.system('play -nq -t alsa synth {} sine {}'.format(duration,f4))
